refactor: create auth folder and transfer routes
