+++ src/routes/[library]/[section]/auth/+layout.svelte
@@ -20,17 +20,17 @@
 		<!-- Logo header -->
 		<div class="items-right flex h-[50px] w-full justify-end gap-4">
 			<img
-				src="../../logos/up-logo.png"
+				src="../../../logos/up-logo.png"
 				class="h-full w-auto"
 				alt="University of the Philippines logo"
 			/>
 			<img
-				src="../../logos/dcs-logo.png"
+				src="../../../logos/dcs-logo.png"
 				class="h-full w-auto"
 				alt="Department of Computer Science logo"
 			/>
 			<img
-				src="../../logos/engglib-logo.png"
+				src="../../../logos/engglib-logo.png"
 				class="h-full w-auto"
 				alt="Engineering Library logo"
 			/>
