+++ src/routes/[library]/[section]/auth/+layout.ts
@@ -1,5 +1,5 @@
 import { error } from '@sveltejs/kit';
-import type { LayoutLoad } from './$types';
+import type { LayoutLoad } from '../$types';
 
 export const load: LayoutLoad = ({ params }) => {
 	if (params.library === 'engglib1') {
@@ -7,7 +7,7 @@ export const load: LayoutLoad = ({ params }) => {
 			libraryName: 'Engineering Library I',
 			libraryBuilding: ' 2/F Melchor Hall, ',
 			libraryStreet: 'S. Osmeña Avenue, UP Diliman',
-			librarySrc: '../../photos/MH.jpg',
+			librarySrc: '../../../photos/MH.jpg',
 			librarySection: params.section
 		};
 	}
@@ -16,7 +16,7 @@ export const load: LayoutLoad = ({ params }) => {
 			libraryName: 'Engineering Library II',
 			libraryBuilding: 'G/F UP Alumni Engineers Centennial Hall,',
 			libraryStreet: 'Velasquez St., UP Diliman',
-			librarySrc: '../../photos/AECH.jpeg',
+			librarySrc: '../../../photos/AECH.jpeg',
 			librarySection: params.section
 		};
 	}
