+++ src/routes/[library]/[section]/auth/register/+page.server.ts
