+++ src/routes/[library]/[section]/auth/register/schema.ts
