+++ src/routes/[library]/[section]/student-dashboard/+page.svelte
@@ -15,18 +15,18 @@
 	import { ServiceStore } from '$lib/stores/ServiceStore';
 	import { ActiveUsageLogStore } from '$lib/stores/UsageLogStore';
 	import type { ServiceFilter, UsageLogFilter } from '$lib/dataTypes/EntityFilters';
-	import { linkRfid } from '../../../supabase/LoginReg';
-	import { readServiceType } from '../../../supabase/ServiceType';
-	import { readService } from '../../../supabase/Service';
-	import { readUsageLog } from '../../../supabase/UsageLog';
-	import { readUser } from '../../../supabase/User';
+	import { linkRfid } from '../auth/supabase/LoginReg';
+	import { readServiceType } from '../auth/supabase/ServiceType';
+	import { readService } from '../auth/supabase/Service';
+	import { readUsageLog } from '../auth/supabase/UsageLog';
+	import { readUser } from '../auth/supabase/User';
 
 	// ----------------------------------------------------------------------------
 	// SESSION FUNCTIONS
 	// ----------------------------------------------------------------------------
 
 	let rfid: string = ''; // rfid linking
-	let library: string = $page.url.pathname.split('/')[1]; // session 
+	let library: string = $page.url.pathname.split('/')[1]; // session
 
 	async function startSession(session: Session | null = null) {
 		// Saves the user's access and refresh tokens in cookies and creates a new session if needed.
@@ -67,14 +67,14 @@
 				toast.error(`Error with creating session: ${error}`);
 				goto('./login');
 			} else {
-                $UserStore.authenticated = true;
-                $UserStore.formData.username = user?.email ? user?.email.split('@')[0] : '';
-                toast.success(`You're now logged in!`);
-            }			
+				$UserStore.authenticated = true;
+				$UserStore.formData.username = user?.email ? user?.email.split('@')[0] : '';
+				toast.success(`You're now logged in!`);
+			}
 		}
-        getUser();
-        getActiveUsageLogs();
-        getServices();
+		getUser();
+		getActiveUsageLogs();
+		getServices();
 		return;
 	}
 
@@ -101,7 +101,7 @@
 	// ----------------------------------------------------------------------------
 
 	async function checkRfidEnter(event: KeyboardEvent) {
-        // checks once RFID has been entered
+		// checks once RFID has been entered
 		if (event.key == 'Enter') {
 			linkRfid(rfid, $UserStore.formData.username);
 		}
@@ -114,7 +114,7 @@
 
 	let isLoggedOut: boolean = false;
 	// const maxSessionDuration: number = 30 * 1000; // seconds * ticks
-    // const reminderTime: number = maxSessionDuration - 5000 // reminds 5 seconds before automatic logout
+	// const reminderTime: number = maxSessionDuration - 5000 // reminds 5 seconds before automatic logout
 	// let logOutReminder = setTimeout(remindLogOut, reminderTime); // user will be reminded of auto logout 5 seconds before
 	// let logOutTimer = setTimeout(logOutUser, maxSessionDuration);
 
@@ -122,8 +122,8 @@
 		// Logs out the user without confirmation and goes to login page
 		await endSession();
 		isLoggedOut = true;
-        // clearTimeout(logOutReminder);
-        // clearTimeout(logOutTimer);
+		// clearTimeout(logOutReminder);
+		// clearTimeout(logOutTimer);
 		goto('./login');
 	}
 
@@ -150,7 +150,7 @@
 	// }
 
 	// function remindLogOut() {
-    //     // Reminds user of automatic logout
+	//     // Reminds user of automatic logout
 	// 	toast(`You will be logged out in 5 seconds unless you select a service.`, {
 	// 		icon: '⏳'
 	// 	});
@@ -161,92 +161,92 @@
 	// READ SERVICES ANG USAGE LOGS
 	// ----------------------------------------------------------------------------
 
-    const serviceFilter: ServiceFilter = {
-        service_type: '',
-        in_use: false,
-        library: library,
-        section: '',
-    }
+	const serviceFilter: ServiceFilter = {
+		service_type: '',
+		in_use: false,
+		library: library,
+		section: ''
+	};
 
-    async function getServices() {
-        // Reads the service types and available services in the current library and section
-        // and puts the returned values in $ServiceStore
-        const { serviceTypes, error } = await readServiceType();
+	async function getServices() {
+		// Reads the service types and available services in the current library and section
+		// and puts the returned values in $ServiceStore
+		const { serviceTypes, error } = await readServiceType();
 
-        if (error) {
-            toast.error(`Error with getting service types: ${error}`);
-            return;
-        } else if (serviceTypes != null) {
-            $ServiceStore.serviceTypes = serviceTypes
-            const { services, error } = await readService(serviceFilter);
+		if (error) {
+			toast.error(`Error with getting service types: ${error}`);
+			return;
+		} else if (serviceTypes != null) {
+			$ServiceStore.serviceTypes = serviceTypes;
+			const { services, error } = await readService(serviceFilter);
 
-            if (error) {
-                toast.error(`Error with getting services: ${error}`);
-                return;
-            } else if (services != null) {
-                $ServiceStore.services = services;
-            }
-        }
+			if (error) {
+				toast.error(`Error with getting services: ${error}`);
+				return;
+			} else if (services != null) {
+				$ServiceStore.services = services;
+			}
+		}
 
-        return;
-    }
+		return;
+	}
 
-    const activeUsageLogFilter: UsageLogFilter = {
-        usagelog_id: 0,
-        start: null,
-        end: null,
-        lib_user_id: parseInt($UserStore.formData.lib_user_id),
-        service_type: '',
-        library: library,
-        section: '',
-    }
+	const activeUsageLogFilter: UsageLogFilter = {
+		usagelog_id: 0,
+		start: null,
+		end: null,
+		lib_user_id: parseInt($UserStore.formData.lib_user_id),
+		service_type: '',
+		library: library,
+		section: ''
+	};
 
-    async function getActiveUsageLogs() {
-        // Reads the active usage logs of the user in the current library and section
-        const { usagelogs, error } = await readUsageLog(activeUsageLogFilter);
+	async function getActiveUsageLogs() {
+		// Reads the active usage logs of the user in the current library and section
+		const { usagelogs, error } = await readUsageLog(activeUsageLogFilter);
 
-        if (error) {
-            toast.error(`Error with getting usagelogs: ${error}`);
-            return;
-        } else if (usagelogs != null) {
-            $ActiveUsageLogStore.activeUsageLogs = usagelogs
-        }
+		if (error) {
+			toast.error(`Error with getting usagelogs: ${error}`);
+			return;
+		} else if (usagelogs != null) {
+			$ActiveUsageLogStore.activeUsageLogs = usagelogs;
+		}
 
-        return;
-    }
+		return;
+	}
 
-    async function getUser(): Promise<boolean> {
-        // gets user information from database
-        const { users, error } = await readUser({
-            lib_user_id: 0,
-            username: $UserStore.formData.username,
-            is_enrolled: null,
-            is_active: null,
-            college: '',
-            program: '',
-            user_type: ''
-        })
-        
-        if (error) {
-            toast.error(`Error with reading user information: ${error}`)
-            return false;
-        } else if (users != null) {
-            $UserStore.formData.lib_user_id = users[0].lib_user_id.toString();
-            $UserStore.formData.college = users[0].college;
-            $UserStore.formData.first_name = users[0].first_name;
-            $UserStore.formData.middle_name = users[0].middle_initial ? users[0].middle_initial : '';
-            $UserStore.formData.last_name = users[0].last_name;
-            $UserStore.formData.user_type = users[0].user_type;
-            $UserStore.formData.college = users[0].college;
-            $UserStore.formData.program = users[0].program ? users[0].program : '';
-        }
+	async function getUser(): Promise<boolean> {
+		// gets user information from database
+		const { users, error } = await readUser({
+			lib_user_id: 0,
+			username: $UserStore.formData.username,
+			is_enrolled: null,
+			is_active: null,
+			college: '',
+			program: '',
+			user_type: ''
+		});
 
-        return true;
-    }
+		if (error) {
+			toast.error(`Error with reading user information: ${error}`);
+			return false;
+		} else if (users != null) {
+			$UserStore.formData.lib_user_id = users[0].lib_user_id.toString();
+			$UserStore.formData.college = users[0].college;
+			$UserStore.formData.first_name = users[0].first_name;
+			$UserStore.formData.middle_name = users[0].middle_initial ? users[0].middle_initial : '';
+			$UserStore.formData.last_name = users[0].last_name;
+			$UserStore.formData.user_type = users[0].user_type;
+			$UserStore.formData.college = users[0].college;
+			$UserStore.formData.program = users[0].program ? users[0].program : '';
+		}
+
+		return true;
+	}
 
 	// ----------------------------------------------------------------------------
 
-    onMount(startSession);
+	onMount(startSession);
 </script>
 
 <Toaster />
